requirements.txt
streamlit>=1.30.0
textstat>=0.7.3
textblob>=0.17.1
PyMuPDF>=1.23.0
python-docx>=1.1.0
openai>=1.10.0
plotly>=5.18.0
nltk>=3.8.1
scikit-learn>=1.4.0
